move FirstOrderMarkovChain into optimizer
